demo1/requirements.txt
Pillow>=8.0.0
pyglet>=1.5.0
